# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all web sessions / 列出所有 Web 会话",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-array_api_SessionItem"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/sessions/{jti}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Kick a web session / 踢掉 Web 会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session jti / 会话标识",
                        "name": "jti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.SessionItem": {
            "type": "object",
            "properties": {
                "idle_timeout_seconds": {
                    "type": "integer"
                },
                "issued_at": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {},
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-array_api_SessionItem": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.SessionItem"
                    }
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                10000,
                10001,
                10002,
                10003,
                10004,
                10005
            ],
            "x-enum-varnames": [
                "CodeOK",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeInternal"
            ]
        }
    }
}